Moved PMA Convert module into qgui.py
