--- qtools2/PmaConvert/setup.py
@@ -1,20 +1,0 @@
-"""
-This is a setup.py script generated by py2applet
-
-Usage:
-    python setup.py py2app
-"""
-
-from setuptools import setup
-
-APP = ['PMA-Convert.py']
-DATA_FILES = ['--iconfile']
-OPTIONS = {'argv_emulation': True,
- 'iconfile': '/Users/joeflack4/projects/PMA-Convert/favicon.icns'}
-
-setup(
-    app=APP,
-    data_files=DATA_FILES,
-    options={'py2app': OPTIONS},
-    setup_requires=['py2app'],
-)
