+++ scripts/pma-convert.py
@@ -1,4 +1,5 @@
-from qtools2.PmaConvert.PmaConvert import run_conversion
+#!/usr/bin/env python
+from qtools2.qgui import run_conversion
 
 if __name__ == '__main__':
     run_conversion()
